+++ package.json
@@ -23,6 +23,7 @@
     "@okta/okta-angular": "1.2.1",
     "body-parser": "^1.19.0",
     "chart.js": "^2.9.4",
+    "component": "^1.1.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "morgan": "^1.10.0",
