+++ src/app/login/login.component.html
@@ -1,44 +1,7 @@
-<div class="container">
 
-  <h1>Login</h1>
-
-  <form id="logInForm">
-    <div [formGroup]="userEmail" class="form-group">
-      <label><b>Username</b></label>
-      <input [class.is-invalid]="userEmail.get('email').invalid" type="email" class="form-control"
-        formControlName="email" placeholder="Enter Username:" id="username" required>
-      <div *ngIf="(getUserEmail().invalid && getUserEmail().touched)|| getUserEmail().dirty">
-        <small *ngIf="getUserEmail().errors?.pattern" class="text-danger">Please enter a valid email address</small>
-      </div>
-      <div *ngIf="getUserEmail().touched|| getUserEmail().dirty">
-        <small *ngIf="valid && !usernameExisted && usernameExisted!=null" class="text-danger">The username doesn't exist. Please retry</small>
-      </div>
-    </div>
-
-    <div [formGroup]="passwords" class="form-group">
-      <label><b>Password</b></label>
-      <input [class.is-invalid]="passwords.get('password').invalid" type="password" class="form-control"
-        formControlName="password" placeholder="Enter Password:" id="password" required>
-
-      <!--Displays this error messages if isn't a valid password-->
-      <div *ngIf="(getUserPassword().invalid && getUserPassword().touched)|| getUserPassword().dirty">
-        <small *ngIf="getUserPassword().errors?.pattern" class="text-danger">Password must contain at least one letter,
-          one symbol, and one number</small>
-      </div>
-      <!-- Displays error message when char length exceeds boundaries -->
-      <div *ngIf="(getUserPassword().invalid && getUserPassword().touched)|| getUserPassword().dirty">
-        <small *ngIf="getUserPassword().errors?.minlength || getUserPassword().errors?.maxlength" class="text-danger">
-          Password must be between 4 and 12 characters</small>
-      </div>
-
-    </div>
-    <!--Displays error message after pressing the login button -->
-    <div *ngIf="signedIn==false && getUserPassword().touched">
-      <small *ngIf="correctPassword==false" class="text-danger">Password Is Incorrect</small>
-    </div>
-
-    <button id="button" type="button" class="btn-primary" (click)="signIn()">Log in</button>
-
-  </form>
-
-</div>
+<input type="text" name="username" id="username" [(ngModel)]="username">
+<br>
+<input type="password" name="password" id="password" [(ngModel)]="password">
+<br>
+<input type="button" value="Login" (click)="Login()">
+<br>
